Split ScoreBoard into folder components
